Add updated schema with choices table; update responses table
This commit adds a table to contain choices for survey questions and updates the responses table to include a column for the choice chosen.

Choices relate to specific questions (1:M question->choices) and the responses table relates to surveys, choices, and contributors.

Note that before adding new responses to the responses table, a SELECT and possible DELETE should be run if the current contributor_id already has responses in the table for the current survey_id.